requirements.txt
# Docker base image: `gcr.io/deeplearning-platform-release/tf2-gpu.2-6:latest`
tensorflow==2.6.2 # The latest should include tensorflow-gpu
tensorflow-datasets==4.4.0
tensorflow-addons==0.15.0
absl-py==0.12.0
gin-config==0.5.0
parameterized==0.8.1
mediapy==1.0.3
scikit-image==0.19.1
apache-beam==2.34.0
